Hack in inline .webm support (working)
Tested with http://techslides.com/demos/sample-videos/small.webm.  The original change almost worked except I tested it on an imgur link, which converts .webm to .gifv.  .gifv is just .webm wrapped in some javascript but it can't be straight embedded in a <video> tag.  The only difference between this and the other PR is the autoplay attribute.